lib.rs for library and main.rs for binary
